pyproject.toml
[project]
name = "agentic-doc-intelligence"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.133.1",
    "langchain-core>=1.2.16",
    "langchain-openai>=0.3.0",
    "langgraph>=1.0.9",
    "mcp>=1.26.0",
    "openai>=2.24.0",
    "pinecone",
    "pdfplumber>=0.11.9",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
    "pillow>=11.1.0",
    "langchain-text-splitters>=1.1.1",
    "langchain-elasticsearch>=1.0.0",
    "langchain-unstructured>=1.0.1",
    "langchain-pinecone>=0.2.13",
    "langgraph-checkpoint-postgres>=3.0.4",
    "langchain-experimental>=0.4.1",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "unstructured[pdf]>=0.21.5",
    "asyncpg>=0.31.0",
    "langchain>=1.2.10",
]
